+++ lib/app/providers/settings_provider.g.dart
@@ -390,6 +390,15 @@ extension UserSettingsSetters on UserSettingsNotifier {
 
   Future<void> setHomeShelves(List<ShelfIdentity> value) =>
       _save(state?.copyWith(homeShelves: value));
+
+  Future<void> setShakeDuringSleepTimer(ShakeDuringSleepTimerAction value) =>
+      _save(state?.copyWith(shakeDuringSleepTimer: value));
+
+  Future<void> setShakeSleepTimerAddMinutes(int value) =>
+      _save(state?.copyWith(shakeSleepTimerAddMinutes: value));
+
+  Future<void> setShakeSensitivity(ShakeSensitivity value) =>
+      _save(state?.copyWith(shakeSensitivity: value));
 }
 
 final currentLibraryProvider = Provider<Library?>(
@@ -906,4 +915,34 @@ final homeShelvesProvider = Provider<List<ShelfIdentity>>(
     ),
   ),
   name: 'homeShelvesProvider',
+);
+
+final shakeDuringSleepTimerProvider = Provider<ShakeDuringSleepTimerAction>(
+  (ref) => ref.watch(
+    userSettingsProvider.select(
+      (s) =>
+          s?.shakeDuringSleepTimer ?? DefaultUserSettings.shakeDuringSleepTimer,
+    ),
+  ),
+  name: 'shakeDuringSleepTimerProvider',
+);
+
+final shakeSleepTimerAddMinutesProvider = Provider<int>(
+  (ref) => ref.watch(
+    userSettingsProvider.select(
+      (s) =>
+          s?.shakeSleepTimerAddMinutes ??
+          DefaultUserSettings.shakeSleepTimerAddMinutes,
+    ),
+  ),
+  name: 'shakeSleepTimerAddMinutesProvider',
+);
+
+final shakeSensitivityProvider = Provider<ShakeSensitivity>(
+  (ref) => ref.watch(
+    userSettingsProvider.select(
+      (s) => s?.shakeSensitivity ?? DefaultUserSettings.shakeSensitivity,
+    ),
+  ),
+  name: 'shakeSensitivityProvider',
 );