+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- shake during sleep timer & sensitivity settings
+
 ### Changed
 
 - replace player slider with new AppSlider supporting cancel-on-pull-up gesture